requirements.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
asgiref==3.4.1
drf-nested-routers==0.93.4
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2